🔧 fix(admin): add product_id to detailFormData in TambahResep page.tsx 🚀 feat(customer): remove 'Rp.' prefix from price display in TransaksiContent 🎉 feat(manajer_operasional): add TableCekBahan component to display stock availability The changes in the TambahResep page add the product_id to the detailFormData when submitting data, ensuring the correct association between ingredients and products. In the TransaksiContent component, the 'Rp.' prefix is removed from the price display for a cleaner UI. The new TableCekBahan component calculates and displays the stock availability of ingredients based on quantity and current stock levels.
🔨 refactor(page.tsx): refactor Home component to HomeMO component for better clarity and consistency
🚀 feat(page.tsx): add functionality to fetch and display transaction data, handle updates, and show invoices
The changes were made to refactor the existing Home component to HomeMO component for better clarity and consistency in naming conventions. Additionally, new functionality was added to fetch transaction data, handle updates, and display invoices based on the fetched data. This enhances the overall functionality and user experience of the page.

🔧 refactor(resep_bahan.tsx): add new interface BahanResep and import related types
ℹ️ chore(resep_bahan.tsx): update export to include BahanResep interface
The changes involve adding a new interface BahanResep and importing related types (Product and Resep) to improve the structure and organization of the code. The export statement is updated to include the newly added BahanResep interface for better clarity and maintainability.